+++ packages/iov-bcp/src/atomicswapmerger.spec.ts
@@ -2,7 +2,7 @@ import { Encoding } from "@iov/encoding";
 
 import { AtomicSwapHelpers } from "./atomicswaphelpers";
 import { AtomicSwapMerger } from "./atomicswapmerger";
-import { ClaimedSwap, OpenSwap, Preimage, SwapState } from "./atomicswaptypes";
+import { ClaimedSwap, OpenSwap, Preimage, SwapProcessState } from "./atomicswaptypes";
 import {
   Address,
   Algorithm,
@@ -35,7 +35,7 @@ describe("AtomicSwapMerger", () => {
     const hash = AtomicSwapHelpers.hashPreimage(preimage);
     const swapId = fromHex("aabbcc") as SwapIdBytes;
     const open: OpenSwap = {
-      kind: SwapState.Open,
+      kind: SwapProcessState.Open,
       data: {
         id: swapId,
         sender: alice,
@@ -63,7 +63,7 @@ describe("AtomicSwapMerger", () => {
     expect(merger.openSwaps().length).toEqual(1);
 
     const expectedSettle: ClaimedSwap = {
-      kind: SwapState.Claimed,
+      kind: SwapProcessState.Claimed,
       data: open.data,
       preimage: preimage,
     };
@@ -86,7 +86,7 @@ describe("AtomicSwapMerger", () => {
     const swapIdA = fromHex("aabbcc") as SwapIdBytes;
     const swapIdB = fromHex("112233") as SwapIdBytes;
     const openA: OpenSwap = {
-      kind: SwapState.Open,
+      kind: SwapProcessState.Open,
       data: {
         id: swapIdA,
         sender: alice,
@@ -97,7 +97,7 @@ describe("AtomicSwapMerger", () => {
       },
     };
     const openB: OpenSwap = {
-      kind: SwapState.Open,
+      kind: SwapProcessState.Open,
       data: {
         id: swapIdB,
         sender: alice,
@@ -138,7 +138,7 @@ describe("AtomicSwapMerger", () => {
     expect(merger.openSwaps().length).toEqual(2);
 
     const expectedSettleB: ClaimedSwap = {
-      kind: SwapState.Claimed,
+      kind: SwapProcessState.Claimed,
       data: openB.data,
       preimage: preimageB,
     };
@@ -146,7 +146,7 @@ describe("AtomicSwapMerger", () => {
     expect(merger.openSwaps().length).toEqual(1);
 
     const expectedSettleA: ClaimedSwap = {
-      kind: SwapState.Claimed,
+      kind: SwapProcessState.Claimed,
       data: openA.data,
       preimage: preimageA,
     };
@@ -161,7 +161,7 @@ describe("AtomicSwapMerger", () => {
     const hash = AtomicSwapHelpers.hashPreimage(preimage);
     const swapId = fromHex("aabbcc") as SwapIdBytes;
     const open: OpenSwap = {
-      kind: SwapState.Open,
+      kind: SwapProcessState.Open,
       data: {
         id: swapId,
         sender: alice,
@@ -191,7 +191,7 @@ describe("AtomicSwapMerger", () => {
     const hash = AtomicSwapHelpers.hashPreimage(preimage);
     const swapId = fromHex("aabbcc") as SwapIdBytes;
     const open: OpenSwap = {
-      kind: SwapState.Open,
+      kind: SwapProcessState.Open,
       data: {
         id: swapId,
         sender: alice,
@@ -219,7 +219,7 @@ describe("AtomicSwapMerger", () => {
     expect(merger.openSwaps().length).toEqual(0);
 
     const expectedSettle: ClaimedSwap = {
-      kind: SwapState.Claimed,
+      kind: SwapProcessState.Claimed,
       data: open.data,
       preimage: preimage,
     };